react client events page
